ai-agent/requirements.txt
openai
python-dotenv
duckduckgo-search
rich
pydantic
pyinstaller
requests
